Add tests for api using mongodb
