+++ tests/utils/scdb.py
@@ -1,0 +1,6 @@
+import shutil
+
+
+def delete_folder(path: str):
+    """Deletes the folder at the given path"""
+    shutil.rmtree(path=path, ignore_errors=True)
